# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service",
        "description": "Service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "local server"
        }
    ],
    "paths": {
        "/api/v1/notifications/send": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Запуск массовой рассылки уведомлений",
                "operationId": "cbff7c8e08218cafea5834936a7d73de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel",
                                    "text",
                                    "user_ids",
                                    "priority"
                                ],
                                "properties": {
                                    "channel": {
                                        "type": "string",
                                        "example": "sms",
                                        "enum": [
                                            "sms",
                                            "email"
                                        ]
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Код: 1234",
                                        "maxLength": 1000
                                    },
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "transactional",
                                        "enum": [
                                            "transactional",
                                            "marketing"
                                        ]
                                    },
                                    "idempotency_key": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "550e8400-e29b-41d4-a716-446655440000",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Уведомления поставлены в очередь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notifications queued"
                                        },
                                        "idempotency_key": {
                                            "type": "string",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        },
                                        "notifications": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "channel": {
                                                        "type": "string"
                                                    },
                                                    "text": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "idempotency_key": {
                                                        "type": "string"
                                                    },
                                                    "gateway_response": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/users/{user_id}/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Получить историю уведомлений пользователя",
                "operationId": "ac7cd4b23fa80cbb6497d203d195ed2c",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список уведомлений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "channel": {
                                                        "type": "string"
                                                    },
                                                    "text": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "idempotency_key": {
                                                        "type": "string"
                                                    },
                                                    "gateway_response": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}